+++ redux-toolkit/userData.js
@@ -1,31 +1,26 @@
 // step 3: create a slice userData.js
 import { createSlice } from "@reduxjs/toolkit";
+import { initialValues } from "../utils/initialValues";
 
 export const userDataSlice = createSlice({
   name: "userData",
   //   define a state
   initialState: {
-    values: {
-      firstName: "",
-      middleName: "",
-      lastName: "",
-      gender: "",
-      dateOfBirth: "",
-      email: "",
-      phoneNumber: "",
-      completeAddress: "",
-    },
+    values: {initialValues},
   },
   //   define a reducer and action
   reducers: {
     addValues: (state, action) => {
       state.values = action.payload;
     },
+    resetValues: (state) => {
+      state.values = initialValues;
+    },
   },
 });
 
 // export action so that components can modify the state
-export const { addValues } = userDataSlice.actions;
+export const { addValues, resetValues } = userDataSlice.actions;
 
 // export reducers to store
 export default userDataSlice.reducer;